Port manageUserState saga test to Jest
